+++ src/fields/choice_i64.cr
@@ -16,7 +16,7 @@ module Fields
     property is_unique : Bool = false
     # Example: [{5, "Title"}, {10, "Title 2"}].
     # Html tag: select
-    property choices : Array(Tuple(Float64, String)) = Array(Tuple(Float64, String)).new
+    property choices : Array(Tuple(Int64, String)) = Array(Tuple(Int64, String)).new
     # To optimize field traversal in the `paladins/check()` method.
     # Hint: It is recommended not to change.
     getter group : UInt32 = 4
@@ -37,7 +37,7 @@ module Fields
     property is_unique : Bool = false
     # Example: [{5, "Title"}, {10, "Title 2"}].
     # Html tag: select multiple
-    property choices : Array(Tuple(Float64, String)) = Array(Tuple(Float64, String)).new
+    property choices : Array(Tuple(Int64, String)) = Array(Tuple(Int64, String)).new
     # To optimize field traversal in the `paladins/check()` method.
     # Hint: It is recommended not to change.
     getter group : UInt32 = 6
@@ -57,7 +57,7 @@ module Fields
     property is_unique : Bool = false
     # Example: [{5, "Title"}, {10, "Title 2"}].
     # Html tag: select
-    property choices : Array(Tuple(Float64, String)) = Array(Tuple(Float64, String)).new
+    property choices : Array(Tuple(Int64, String)) = Array(Tuple(Int64, String)).new
     # To optimize field traversal in the `paladins/check()` method.
     # Hint: It is recommended not to change.
     getter group : UInt32 = 5
@@ -77,7 +77,7 @@ module Fields
     property is_unique : Bool = false
     # Example: [{5, "Title"}, {10, "Title 2"}].
     # Html tag: select
-    property choices : Array(Tuple(Float64, String)) = Array(Tuple(Float64, String)).new
+    property choices : Array(Tuple(Int64, String)) = Array(Tuple(Int64, String)).new
     # To optimize field traversal in the `paladins/check()` method.
     # Hint: It is recommended not to change.
     getter group : UInt32 = 7
